requirements.txt
# Основные зависимости веб-приложения
# WSGI сервер для запуска Flask
gunicorn>=20.0.4
# Веб-фреймворк
Flask>=2.3.3
# HTTP-запросы для клиента
requests>=2.23.0
# Интеграция Bootstrap с Flask
Flask-Bootstrap>=3.3.7.1
# Формы и валидация
Flask-WTF>=1.2.1
# Обработка изображений
Pillow>=9.5.0 --only-binary=all
# Парсинг XML/XSLT
lxml>=4.9.4
# Нейросети (CPU версия)
tensorflow-cpu>=2.13.0
# Высокоуровневый API для нейросетей
keras>=2.13.1

# Дополнительные зависимости для обработки изображений
# Компьютерное зрение (без GUI)
opencv-python-headless>=4.8.1.78
# Научные вычисления
numpy>=1.24.3
# Построение графиков
matplotlib>=3.7.2
# Алгоритмы обработки изображений
scikit-image>=0.21.0
